multer changes

+++ routes/users.js
@@ -16,7 +16,7 @@ router.put("/:id", async (req, res) => {
         {
           $set: req.body,
         },
-        { new: true }
+        { returnDocument: "after" }
       );
       res.status(200).json(updatedUser);
     } catch (err) {
@@ -29,33 +29,42 @@ router.put("/:id", async (req, res) => {
 
 //DELETE
 router.delete("/:id", async (req, res) => {
+  console.log("delete in process");
   if (req.body.userId === req.params.id) {
     try {
       const user = await User.findById(req.params.id);
-      try {
-        await Post.deleteMany({ username: user.username });
-        await User.findByIdAndDelete(req.params.id);
-        res.status(200).json("User has been deleted...");
-      } catch (err) {
-        res.status(500).json(err);
+      if (user) {
+        try {
+          //delete user's all posts
+          await Post.deleteMany({ username: user.username });
+          // delete user
+          await User.findByIdAndDelete(req.params.id);
+          res.status(200).send({ Message: "user deleted Successfully", user });
+        } catch (err) {
+          console.log("Error in delete");
+          res.status(500).send(err);
+        }
       }
     } catch (err) {
-      res.status(404).json("User not found!");
+      console.log("err", err);
+      res.status(404).send("user not found");
     }
   } else {
-    res.status(401).json("You can delete only your account!");
+    res.status(401).send("Youre not permitted to delete others posts");
   }
 });
+
 
 //GET USER
 router.get("/:id", async (req, res) => {
+  console.log("user getting process");
   try {
     const user = await User.findById(req.params.id);
+    console.log(user);
     const { password, ...others } = user._doc;
-    res.status(200).json(others);
-  } catch (err) {
-    res.status(500).json(err);
+    res.status(200).send({ others });
+  } catch (error) {
+    res.status(500).send("Error in getting user");
   }
 });
-
 module.exports = router;
